more libraries

+++ CSharp-Core/src/prompto/type/DecimalType.cs
@@ -1,7 +1,6 @@
 using System;
 using prompto.runtime;
 using Decimal = prompto.value.Decimal;
-using System.Collections.Generic;
 using prompto.value;
 using prompto.store;
 
@@ -30,7 +29,7 @@ namespace prompto.type
         
         public override Type ToCSharpType()
         {
-            return typeof(Decimal);
+			return typeof(Double);
         }
